requirements.txt
google-genai==1.20.0
google-generativeai==0.8.5
requests==2.32.3
monica-client==1.0.0a1
gradio==4.28.0